Initial Commit
After working on this throughout the day in short bursts, I've decide to
make an online repo for safekeeping as well as for record. This commit
includes a very incomplete GUI, and several incomplete classes for the
simulator itself. As far as classes go, most of the source files are
created, but need to be written.