Considering rewrite
As of late, this code has not embraced OOP so much as I had anticipated.
The GUI file is expected to balloon with repetitive functions, and not
everything in place seems sustainable. At the same time, this experience
with Java GUI functions is invaluable. As far as actual changes go, I
added some listeners and actions to the torrent creator spinners.