Refactoring - PeerCreator completeq
Moved the PeerCreator tab to a new class and reorganized code. All went
well, and the program works as expected. Also added an update method
that updates the JList on tab changes. Will need to optimize in the
future, but it will work for the time being. Also looking forward to
adding a scrollbar for the PeerList.