Major Overhaul
Changed many classes fundamentally, having many types carry static lists
containing all existing objects. Still trying to wrap my head around how
that works on a low-level. I decentralized almost everything, and the
Sim doesn't do much except act as a buffer for ticks. The current build
doesn't produce results yet, but I should find what the holdup is soon.
I eliminated much of the support for color, as it just got in the way
and is not a need-to-have. My goal is to have a working simulation, and
will go from there with extra features and other support.